Update elasticsearch-operator_OperatorGroup.yaml
Reference: https://access.redhat.com/solutions/6459071


    Operator installation/upgrade fails stating:
    Raw

    Bundle unpacking failed. Reason: DeadlineExceeded, and Message: Job was active longer than specified deadline".

Resolution

This issue has been reported to Red Hat engineering. It is being tracked in Bug OCPBUGS-6771.

With the release of OCP 4.14.14 and 4.15.0 we have released the core functionality to fix the "bundle unpacking failed. Reason: DeadlineExceeded".
There is an additional annotation to the operatorGroup for every to be installed operator needs to be added before applying it as follows: operatorframework.io/bundle-unpack-min-retry-interval